job-template/job/AccessMgtServer/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic>=2.0.0
PyYAML>=6.0
pandas>=2.0.0
numpy>=1.24.0
numexpr>=2.8.0
python-docx>=1.1.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-multipart>=0.0.6

